move variant split_alt to binf module
